Move compound types into packages/math
