--- src/lib/compounds.ts
@@ -1,3 +1,0 @@
-import type { Position, Rect, Size } from 'packages/ui/compounds';
-
-export type { Position, Rect, Size };
