+++ src/packages/grid/spaces/interface.ts
@@ -1,4 +1,4 @@
-import type { Position } from '$lib/compounds';
+import type { Position } from 'packages/math';
 
 export type PositionedSquare = {
 	center: Position;
