+++ src/routes/games/[game]/board/tokens/token-management.ts
@@ -1,7 +1,7 @@
 import type { ClientSocket } from '$lib/client/communication';
 import { Board } from '$lib/client/state';
 import type { GetPayload, TokenSnippet } from '$lib/net';
-import type { Position } from 'packages/ui/compounds';
+import type { Position } from 'packages/math';
 import type { Direction } from 'packages/undo-redo/action';
 import { type HistoryStore } from 'packages/undo-redo/history';
